libvirt: set USERADD_DEPENDS to avoid package_write_rpm failure
Issue: LIN1024-5015

When do_package uses state cache and is not re-run but do_package_write_rpm
is re-run, then we get the following error:

  Exception: KeyError: 'getpwuid(): uid not found: 996'

This is because libvirt does chown to polkitd in do_install and polkitd
is from polkit.

Here's an easy way to reproduce this issue:

  bitbake libvirt -c cleansstate && bitbake libvirt -c package && \
  bitbake libvirt -c clean && bitbake libvirt -c package_write_rpm

So make use of USERADD_DEPENDS to ensure polkitd exists to fix this issue.

(LOCAL REV: NOT UPSTREAM) -- sent to meta-virt on 20241108

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>